requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface==0.0.95
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
pandas==2.1.3
gdown==4.7.1
tqdm==4.66.1

